requirements.txt
streamlit
google-generativeai
markdown
python-docx
beautifulsoup4
PyPDF2